feat: support login with id token (#406)
* feat: support login with id token

Adding support for id token login to the provider configuration.
The id token replaces username and password and is passed to a new
special login endpoint of the btp CLI server.
The resulting session id is read from a response header. Currently,
the response does not contain any authentication info, i.e. the
terraform provider does not support the whoami datasource when
logging in via id token.