feat: improve id token login
This is a follow up for PR #406 adapting latest changes to the
idtoken login endpoint in the btp CLI server.

The idtoken based login is now providing authentication information
and is therefore compatible again with the btp_whoami data source.

Furthermore, the error handling on server side was improved which
is resulting in more helpful error messages for the terraform user.